Move domain::Tags::FunctionsOfTimeInitialize to DomainCreator library
The base tag is left in the original file. Only the simple tag was moved
