+++ tests/Unit/ControlSystem/Test_UpdateFunctionOfTime.cpp
@@ -11,6 +11,7 @@
 
 #include "ControlSystem/UpdateFunctionOfTime.hpp"
 #include "DataStructures/DataVector.hpp"
+#include "Domain/Creators/Tags/FunctionsOfTime.hpp"
 #include "Domain/FunctionsOfTime/FunctionOfTime.hpp"
 #include "Domain/FunctionsOfTime/PiecewisePolynomial.hpp"
 #include "Domain/FunctionsOfTime/QuaternionFunctionOfTime.hpp"
