+++ tests/Unit/Domain/FunctionsOfTime/Test_Tags.cpp
@@ -3,35 +3,11 @@
 
 #include "Framework/TestingFramework.hpp"
 
-#include "Domain/Creators/OptionTags.hpp"
-#include "Domain/FunctionsOfTime/FunctionOfTime.hpp"
 #include "Domain/FunctionsOfTime/Tags.hpp"
 #include "Helpers/DataStructures/DataBox/TestHelpers.hpp"
-#include "Utilities/TypeTraits.hpp"
 
 namespace domain {
-
-template <bool Override>
-struct Metavariables {
-  static constexpr size_t volume_dim = 3;
-  static constexpr bool override_functions_of_time = Override;
-};
-
 SPECTRE_TEST_CASE("Unit.Domain.FunctionsOfTime.Tags", "[Domain][Unit]") {
   TestHelpers::db::test_base_tag<Tags::FunctionsOfTime>("FunctionsOfTime");
-  TestHelpers::db::test_simple_tag<Tags::FunctionsOfTimeInitialize>(
-      "FunctionsOfTime");
-
-  CHECK(std::is_same_v<
-        Tags::FunctionsOfTimeInitialize::option_tags<Metavariables<true>>,
-        tmpl::list<
-            domain::OptionTags::DomainCreator<Metavariables<true>::volume_dim>,
-            domain::FunctionsOfTime::OptionTags::FunctionOfTimeFile,
-            domain::FunctionsOfTime::OptionTags::FunctionOfTimeNameMap>>);
-
-  CHECK(std::is_same_v<
-        Tags::FunctionsOfTimeInitialize::option_tags<Metavariables<false>>,
-        tmpl::list<domain::OptionTags::DomainCreator<
-            Metavariables<false>::volume_dim>>>);
 }
 }  // namespace domain
