+++ tests/Unit/Helpers/IO/Observers/ObserverHelpers.hpp
@@ -6,8 +6,8 @@
 #include <functional>
 
 #include "Domain/Creators/Tags/Domain.hpp"
+#include "Domain/Creators/Tags/FunctionsOfTime.hpp"
 #include "Domain/Domain.hpp"
-#include "Domain/FunctionsOfTime/Tags.hpp"
 #include "Domain/Structure/ElementId.hpp"
 #include "Framework/ActionTesting.hpp"
 #include "IO/Observer/Helpers.hpp"
