+++ tests/Unit/IO/Observers/Test_ReductionObserver.cpp
@@ -13,6 +13,7 @@
 
 #include "DataStructures/Matrix.hpp"
 #include "Domain/Creators/Tags/Domain.hpp"
+#include "Domain/Creators/Tags/FunctionsOfTime.hpp"
 #include "Domain/Domain.hpp"
 #include "Domain/FunctionsOfTime/Tags.hpp"
 #include "Domain/Structure/ElementId.hpp"
